Added option to fix air vehicle goal pose height above ground level

+++ FastMarching3D_ROS/msfm3d_node.cpp
@@ -1979,7 +1979,7 @@ int main(int argc, char **argv)
   bool fixGoalHeightAGL;
   float goalHeightAGL;
   n.param("msfm3d/fixGoalHeightAGL", fixGoalHeightAGL, false);
-  n.param("msfm3d/goalHeightAGL", goalHeightAGL, 0.64);
+  n.param("msfm3d/goalHeightAGL", goalHeightAGL, ((float)0.64);
   planner.fixGoalHeightAGL = fixGoalHeightAGL;
   planner.goalHeightAGL = goalHeightAGL;
